multimodal_parser/requirements.txt
openai
pydantic
PyMuPDF
Pillow
dashscope
